+++ src/schemathesis/utils.py
@@ -284,10 +284,13 @@ def traverse_schema(schema: Schema, callback: Callable[..., Dict[str, Any]], *ar
     return schema
 
 
-def _warn_deprecation(*, thing: str, removed_in: str, replacement: str) -> None:
+def warn_filtration_arguments(name: str) -> None:
+    warn_deprecation(thing=f"Argument `{name}`", removed_in="4.0", replacement="`include` or `exclude` methods")
+
+
+def warn_deprecation(*, thing: str, removed_in: str, replacement: str) -> None:
     warnings.warn(
-        f"Property `{thing}` is deprecated and will be removed in Schemathesis {removed_in}. "
-        f"Use `{replacement}` instead.",
+        f"{thing} is deprecated and will be removed in Schemathesis {removed_in}. " f"Use {replacement} instead.",
         DeprecationWarning,
     )
 
@@ -296,7 +299,7 @@ def deprecated_property(*, removed_in: str, replacement: str) -> Callable:
     def wrapper(prop: Callable) -> Callable:
         @property  # type: ignore
         def inner(self: Any) -> Any:
-            _warn_deprecation(thing=prop.__name__, removed_in=removed_in, replacement=replacement)
+            warn_deprecation(thing=f"Property `{prop.__name__}`", removed_in=removed_in, replacement=replacement)
             return prop(self)
 
         return inner
@@ -307,7 +310,7 @@ def deprecated_property(*, removed_in: str, replacement: str) -> Callable:
 def deprecated(*, removed_in: str, replacement: str) -> Callable:
     def wrapper(func: Callable) -> Callable:
         def inner(*args: Any, **kwargs: Any) -> Any:
-            _warn_deprecation(thing=func.__name__, removed_in=removed_in, replacement=replacement)
+            warn_deprecation(thing=f"Function `{func.__name__}`", removed_in=removed_in, replacement=replacement)
             return func(*args, **kwargs)
 
         return inner